+++ src/Symfony/Component/PropertyAccess/Tests/Fixtures/TraversableArrayObject.php
@@ -15,7 +15,7 @@ namespace Symfony\Component\PropertyAccess\Tests\Fixtures;
  * This class is a hand written simplified version of PHP native `ArrayObject`
  * class, to show that it behaves differently than the PHP native implementation.
  */
-class TraversableArrayObject implements \ArrayAccess, \IteratorAggregate, \Countable, \Serializable
+class TraversableArrayObject implements \ArrayAccess, \IteratorAggregate, \Countable
 {
     private $array;
 
@@ -63,18 +63,8 @@ class TraversableArrayObject implements \ArrayAccess, \IteratorAggregate, \Count
         return $this->array;
     }
 
-    public function serialize(): string
-    {
-        return serialize($this->__serialize());
-    }
-
     public function __unserialize(array $data): void
     {
         $this->array = $data;
-    }
-
-    public function unserialize($serialized)
-    {
-        $this->__unserialize((array) unserialize((string) $serialized));
     }
 }